AI/requirements.txt
# Core
google-adk>=1.3.0
google-genai>=0.4.0
python-dotenv>=1.0.1
colorama>=0.4.6
tqdm>=4.66.4

# RAG (LangChain helpers + FAISS wrapper)
pypdf>=5.4.0
langchain>=0.3.15
langchain-community>=0.3.20
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
numpy>=1.26
pandas>=2.1.4
openpyxl>=3.1.5

# Utilities
typer>=0.12.3
requests==2.32.3

#google-cloud-storage>=3.0.0
